+++ .idea/gradle.xml
@@ -13,8 +13,8 @@
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/core" />
             <option value="$PROJECT_DIR$/feature" />
-            <option value="$PROJECT_DIR$/feature-others" />
             <option value="$PROJECT_DIR$/feature/anime" />
+            <option value="$PROJECT_DIR$/feature/others" />
           </set>
         </option>
       </GradleProjectSettings>
