+++ app/build.gradle
@@ -18,7 +18,7 @@ android {
         minSdk 21
         targetSdk 33
         versionCode 1
-        versionName "1.1.1"
+        versionName "1.2.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -44,8 +44,7 @@ android {
 dependencies {
     implementation project(':core')
     implementation project(':feature:anime')
-//    implementation project(':feature-anime')
-    implementation project(':feature-others')
+    implementation project(':feature:others')
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 }
