--- app/src/main/java/com/lelestacia/lelenimexml/view/gallery/GalleryViewModel.kt
@@ -1,13 +1,0 @@
-package com.lelestacia.lelenimexml.view.gallery
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class GalleryViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is gallery Fragment"
-    }
-    val text: LiveData<String> = _text
-}
