+++ feature/anime/src/main/res/navigation/anime_navigation.xml
@@ -20,10 +20,10 @@
         <action
             android:id="@+id/anime_to_history"
             app:destination="@id/historyFragment"
-            app:enterAnim="@android:anim/slide_in_left"
-            app:exitAnim="@android:anim/slide_out_right"
-            app:popEnterAnim="@android:anim/slide_in_left"
-            app:popExitAnim="@android:anim/slide_out_right" />
+            app:enterAnim="@anim/left_to_center"
+            app:exitAnim="@anim/center_to_right"
+            app:popEnterAnim="@anim/right_to_center"
+            app:popExitAnim="@anim/center_to_left" />
     </fragment>
 
     <fragment
