+++ feature/others/src/main/res/layout/item_tv.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<TextView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/tv_text"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:background="@color/white"
+    android:textAppearance="@style/TextAppearance.Material3.BodyMedium"
+    android:padding="8dp"
+    android:textColor="@color/black"
+    tools:text="Jambi" />
