+++ feature/others/src/main/res/navigation/settings_navigation.xml
@@ -3,11 +3,11 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/settings_navigation"
-    app:startDestination="@id/nav_settings">
+    app:startDestination="@id/settingsFragment">
 
     <fragment
-        android:id="@+id/nav_settings"
+        android:id="@+id/settingsFragment"
         android:name="com.lelestacia.lelenimexml.feature.others.SettingsFragment"
-        android:label="Settings"
+        android:label="fragment_settings"
         tools:layout="@layout/fragment_settings" />
 </navigation>
